+++ pcs/app.py
@@ -278,6 +278,9 @@ def main(argv=None):
             utils.get_library_wrapper(), argv, utils.get_input_modifiers()
         )
     except LibraryError as e:
+        if e.output:
+            sys.stderr.write(e.output)
+            sys.exit(1)
         process_library_reports(e.args)
     except errors.CmdLineInputError:
         if argv and argv[0] in cmd_map: