+++ pcs/lib/errors.py
@@ -1,2 +1,11 @@
+from typing import Optional
+
+
 class LibraryError(Exception):
-    pass
+    def __init__(self, *args, output: Optional[str] = None):
+        super().__init__(*args)
+        self._output = output
+
+    @property
+    def output(self):
+        return self._output