+++ pcs/pcs_internal.py
@@ -177,6 +177,7 @@ def main() -> None:
             report_list=(
                 cli_env.report_processor.processed_items + list(e.args)
             ),
+            data=e.output,
         )
     except json.JSONDecodeError as e:
         _exit(