+++ CHANGELOG.md
@@ -4,12 +4,15 @@
 
 ### Changed
 - Pcs no longer depends on python3-distro package
+- 'pcs status xml' now prints cluster status in the new format provided by
+  Pacemaker 2.1 ([rhbz#1985981])
 
 ### Removed
 - Deprecated obsolete commands `pcs config import-cman` and `pcs config export
   pcs-commands|pcs-commands-verbose` have been removed ([rhbz#1881064])
 
 [rhbz#1881064]: https://bugzilla.redhat.com/show_bug.cgi?id=1881064
+[rhbz#1985981]: https://bugzilla.redhat.com/show_bug.cgi?id=1985981
 
 
 ## [0.10.10] - 2021-08-19